rename to Modal and remove old css
